webhook-ingester/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pulsar-client==3.3.0
pydantic==2.4.2
google-cloud-pubsub==2.21.1